Fix circular depedendecy
